+++ publish/index.js
@@ -13,8 +13,9 @@ require('./src/commands/extract-staking-balances').cmd(program);
 require('./src/commands/import-fee-periods').cmd(program);
 require('./src/commands/migrate-binary-option-markets').cmd(program);
 require('./src/commands/nominate').cmd(program);
-require('./src/commands/relayed-nominate').cmd(program);
+require('./src/commands/nominate-relay').cmd(program);
 require('./src/commands/owner').cmd(program);
+require('./src/commands/owner-relay').cmd(program);
 require('./src/commands/persist-tokens').cmd(program);
 require('./src/commands/prepare-deploy-detect-diff').cmd(program);
 require('./src/commands/prepare-deploy').cmd(program);
@@ -27,6 +28,5 @@ require('./src/commands/verify').cmd(program);
 require('./src/commands/versions-history').cmd(program);
 require('./src/commands/versions-update').cmd(program);
 require('./src/commands/deploy-shorting-rewards').cmd(program);
-require('./src/commands/accept-relay-owner').cmd(program);
 
 program.parse(process.argv);
